Fix more possible race conditions when removing measurements

+++ backend/configuration/src/integrationTest/java/com/elvaco/mvp/web/DashboardControllerTest.java
@@ -51,7 +51,9 @@ public class DashboardControllerTest extends IntegrationTest {
 
   @After
   public void tearDown() {
-    measurementJpaRepository.deleteAll();
+    if (isPostgresDialect()) {
+      measurementJpaRepository.deleteAll();
+    }
   }
 
   @Test
